+++ app/src/main/java/com/example/cosc341_project/ui/feed/FeedViewModel.java
@@ -1,16 +1,16 @@
-package com.example.cosc341_project.ui.home;
+package com.example.cosc341_project.ui.feed;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class FeedViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public FeedViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is feed fragment");
     }
 
     public LiveData<String> getText() {
