+++ app/src/main/res/layout/fragment_feed.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.feed.FeedFragment">
 
     <TextView
-        android:id="@+id/text_home"
+        android:id="@+id/text_feed"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
