+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">Cosc341_Project</string>
 
-    <string name="title_home">Home</string>
+    <string name="title_feed">Feed</string>
     <string name="title_map">Map</string>
     <string name="title_post">Post</string>
     <string name="title_profile">Profile</string>
